requirements.txt
# Core packages
flask==3.1.1
python-dotenv==1.1.0

# LangChain components
langchain==0.3.26
langchain-community==0.3.26
langchain-pinecone==0.2.8

# Hugging Face Inference API support
huggingface_hub==0.23.1
transformers==4.42.2  # required to format prompts, etc., even if using API
requests==2.31.0      # for calling the HF API

# Embeddings
sentence-transformers==4.1.0  # if you're using local embedding or HF models via sentence-transformers

# Document parsing
pypdf==5.6.1


